Add files via upload
This is a tarball of the game's main source code. Please note that the underlying code base is a ROM2.4 MUD, which was built on MERC MUD, which was built on DikuMUD.  You are subject to all of their licenses as well as the MIT.  Unfortunately I don't have a copy of the original licenses any more, but you can find them by Googling around, I'm sure.  I included a few copies I was able to find for reference.